Update conditional_statements_js
In JavaScript, conditional statements allow you to execute different blocks of code based on whether a condition is true or false. If the condition evaluates to true, the code within the if block will be executed. If the condition evaluates to false, the code within the else block will be executed. Here are some examples of conditional statements in JavaScript: